adk-a2a-chatbot/pyproject.toml
[project]
name = "adk-chatbot"
version = "0.1.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.0",
    "amadeus>=12.0.0",
    "click>=8.2.1",
    "dotenv>=0.9.9",
    "google-adk[eval]>=1.9.0",
    "google-cloud-aiplatform>=1.101.0",
    "google-genai>=1.28.0",
    "langchain-community>=0.3.27",
    "mysql-connector-python>=9.4.0",
    "pinecone>=7.3.0",
    "requests>=2.32.4",
    "tavily-python>=0.7.10",
    "uvicorn>=0.35.0",
]
